\name{generate.MVN}
\alias{generate.MVN}
\title{extracts the free parameters and the covariance matrix to generate samples using a multivariate normal distribution}
\description{
Creates an R object containing the header information for a file, and a matrix of multivariate samples.
}
\usage{
generate.MVN(file,path,output.file,sample.size=1,step=1,method="chol")
}
\arguments{
  \item{file}{character string giving the file name of a standard CASAL output file (i.e., containing the covariance matrix and MPD estimates).}
  \item{path}{Optional character string giving the path to 'file'.}
  \item{output.file}{character string giving the file name to oputput the random samples.}
  \item{sample.size}{Number of multivariate samples to generate.}
  \item{step}{Controls the size of the vector difference between the point estimate in mpdfile and the new starting point.}
  \item{method}{Matrix decomposition used to determine the matrix root of the covariance matrix, possible methods are singular value decomposition ("svd", default) and Cholesky decomposition ("chol").}
}
\details{
  Generates a set of parameter vectors generated from a point estimate and a covariance matrix.
  Each new random sample is generated by stepping randomly away from the existing point estimate in a direction controlled by the estimated covariance matrix (and step) at that point. Any parameters which are found to be outside the parameter bounds are shifted to half-way between the existing estimate and the bound that was exceeded. The random sample is generated as a random vector from a multivariate normal distribution with mean at the origin and covariance matrix equal to that estimated at the existing point estimate, nultiplied by step.
}
\value{
  A list object.
  generate.MVN returns an invisible list containing the items
  \item{header}{the text of a header line typically found in a samples file from an MCMC.}
  \item{MVN}{a matrix of randomly generated samples from the multivariate normal, with columns representing the free parameters and a row for each randomly generated sample.}
}
\note{
Illegal names can be used to name free parameters in CASAL. Such names are converted to legal names by enclosing the string in quotes. The exact format of the object on the file is subject to change. No error checking is undertaken to ensure that the data in the R object accurately represents the data in the CASAL file. Note also that exact equality of read-in numeric data is machine specific.
}
\seealso{
\code{\link{casal-package}}
}
