\name{generate.MVU}
\alias{generate.MVU}
\title{extracts the free parameters to generate samples using a multivariate uniform distribution}
\description{
Creates an R object containing the header information for a file, and a matrix of random samples.
}
\usage{
generate.MVU(file,path,output.file,sample.size=1)
}
\arguments{
  \item{file}{character string giving the file name of a standard CASAL output file (i.e., containing the MPD estimates).}
  \item{path}{Optional character string giving the path to 'file'.}
  \item{output.file}{character string giving the file name to oputput the random samples.}
  \item{sample.size}{Number of samples to generate.}
}
\details{
  Generates a set of parameter vectors generated between the parameter bounds.
  Each new random sample is generated by a random uniform distribution between the parameter bounds. The random sample is generated as a random vector from a multivariate uniform distribution.
}
\value{
  A list object.
  generate.MVU returns an invisible list containing the items
  \item{header}{the text of a header line typically found in a samples file from an MCMC.}
  \item{MVU}{a matrix of randomly generated samples, with columns representing the free parameters and a row for each randomly generated sample.}
}
\note{
Illegal names can be used to name free parameters in CASAL. Such names are converted to legal names by enclosing the string in quotes. The exact format of the object on the file is subject to change. No error checking is undertaken to ensure that the data in the R object accurately represents the data in the CASAL file. Note also that exact equality of read-in numeric data is machine specific.
}
\seealso{
\code{\link{casal-package}}
}
